labs/01_chatbot_langchain/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
streamlit
chromadb
tiktoken>=0.5.0 